move CLI to interrupt based
